Add Modelio Module Template
Added the Modelio Module template to the repository. The template includes the necessary files and directory structure for developing Modelio Modules. This commit helps to initialize the project with a starting point for further development.
The template can be set up by following the instructions in the 'Module Developer Guide' section of the Modelio documentation on Github.